2.8.0 1、本次升级一定要清掉redis中所有缓存 2、统一下单接口中传微信的商户订单号由orderId换成orderNo（商城版特有） 3、完善swagger文档 4、后台增加订单取消功能（商城版特有） 5、退款订单不能删除（商城版特有） 6、hutool升级到5.2.2版本 7、spring-boot升级到2.2.5.RELEASE版本 8、spring-boot-admin升级到2.2.2版本 9、spring-cloud升级到Hoxton.SR3版本 10、spring-cloud-alibaba升级到2.2.0.RELEASE版本 11、mybatis-plus升级到3.3.1版本 12、增加订单改价功能（商城版特有） 13、新增拼团功能（商城版特有） 14、`base_mall`.`order_info`新增`market_id`字段记录营销id（砍价ID、拼团ID）（商城版特有） 15、Avue升级到2.4.1版本 16、购物车全选问题https://git.joolun.com/joolun-mp-ma/mp-ma/issues/126（商城版特有） 17、小程序增加倒计时组件（商城版特有）

+++ joolun/base-mall/base-mall-admin/src/main/java/com/joolun/cloud/mall/admin/controller/BargainInfoController.java
@@ -12,8 +12,12 @@ import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.joolun.cloud.common.core.util.R;
 import com.joolun.cloud.common.log.annotation.SysLog;
+import com.joolun.cloud.mall.admin.service.BargainUserService;
+import com.joolun.cloud.mall.common.constant.MyReturnCode;
 import com.joolun.cloud.mall.common.entity.BargainInfo;
 import com.joolun.cloud.mall.admin.service.BargainInfoService;
+import com.joolun.cloud.mall.common.entity.BargainUser;
+import io.swagger.annotations.ApiOperation;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
@@ -34,6 +38,7 @@ import io.swagger.annotations.Api;
 public class BargainInfoController {
 
     private final BargainInfoService bargainInfoService;
+	private final BargainUserService bargainUserService;
 
     /**
      * 分页列表
@@ -41,8 +46,9 @@ public class BargainInfoController {
      * @param bargainInfo 砍价
      * @return
      */
+	@ApiOperation(value = "分页列表")
     @GetMapping("/page")
-    @PreAuthorize("@ato.hasAuthority('mall_bargaininfo_index')")
+    @PreAuthorize("@ato.hasAuthority('mall:bargaininfo:index')")
     public R getPage(Page page, BargainInfo bargainInfo) {
         return R.ok(bargainInfoService.page(page, Wrappers.query(bargainInfo)));
     }
@@ -52,8 +58,9 @@ public class BargainInfoController {
 	 * @param bargainInfo 砍价
 	 * @return
 	 */
+	@ApiOperation(value = "list列表")
 	@GetMapping("/list")
-	@PreAuthorize("@ato.hasAuthority('mall_bargaininfo_index')")
+	@PreAuthorize("@ato.hasAuthority('mall:bargaininfo:index')")
 	public R getList(BargainInfo bargainInfo) {
 		return R.ok(bargainInfoService.list(Wrappers.query(bargainInfo)));
 	}
@@ -63,8 +70,9 @@ public class BargainInfoController {
      * @param id
      * @return R
      */
+	@ApiOperation(value = "砍价查询")
     @GetMapping("/{id}")
-    @PreAuthorize("@ato.hasAuthority('mall_bargaininfo_get')")
+    @PreAuthorize("@ato.hasAuthority('mall:bargaininfo:get')")
     public R getById(@PathVariable("id") String id) {
         return R.ok(bargainInfoService.getById(id));
     }
@@ -74,9 +82,10 @@ public class BargainInfoController {
      * @param bargainInfo 砍价
      * @return R
      */
+	@ApiOperation(value = "砍价新增")
     @SysLog("新增砍价")
     @PostMapping
-    @PreAuthorize("@ato.hasAuthority('mall_bargaininfo_add')")
+    @PreAuthorize("@ato.hasAuthority('mall:bargaininfo:add')")
     public R save(@RequestBody BargainInfo bargainInfo) {
         return R.ok(bargainInfoService.save(bargainInfo));
     }
@@ -86,9 +95,10 @@ public class BargainInfoController {
      * @param bargainInfo 砍价
      * @return R
      */
+	@ApiOperation(value = "砍价修改")
     @SysLog("修改砍价")
     @PutMapping
-    @PreAuthorize("@ato.hasAuthority('mall_bargaininfo_edit')")
+    @PreAuthorize("@ato.hasAuthority('mall:bargaininfo:edit')")
     public R updateById(@RequestBody BargainInfo bargainInfo) {
         return R.ok(bargainInfoService.updateById(bargainInfo));
     }
@@ -98,10 +108,16 @@ public class BargainInfoController {
      * @param id
      * @return R
      */
+	@ApiOperation(value = "砍价删除")
     @SysLog("删除砍价")
     @DeleteMapping("/{id}")
-    @PreAuthorize("@ato.hasAuthority('mall_bargaininfo_del')")
+    @PreAuthorize("@ato.hasAuthority('mall:bargaininfo:del')")
     public R removeById(@PathVariable String id) {
+		int count = bargainUserService.count(Wrappers.<BargainUser>lambdaQuery()
+				.eq(BargainUser::getBargainId,id));
+		if(count > 0){
+			return R.failed(MyReturnCode.ERR_80020.getMsg());
+		}
         return R.ok(bargainInfoService.removeById(id));
     }
